+++ packages/article-skeleton/src/article-skeleton.js
@@ -1,9 +1,6 @@
 import React, { Fragment, useRef, useEffect, useState } from "react";
 import PropTypes from "prop-types";
-import {
-  CanShowPuzzleSidebar,
-  getDomainSpecificUrl
-} from "@times-components/utils";
+import { CanShowPuzzleSidebar } from "@times-components/utils";
 import { AdContainer } from "@times-components/ad";
 import ArticleExtras from "@times-components/article-extras";
 import LazyLoad from "@times-components/lazy-load";
@@ -140,7 +137,8 @@ const ArticleSkeleton = ({
       window.removeEventListener("scroll", handleScroll);
     };
   }, []);
-  const { hostName, canonicalUrl, breadcrumbs } = articleDataFromRender || {};
+  const { hostName, canonicalUrl, breadcrumbs, getDomainSpecificUrl } =
+    articleDataFromRender || {};
   const articleUrl =
     hostName && canonicalUrl ? `${hostName}${canonicalUrl}` : url;
 
@@ -178,9 +176,10 @@ const ArticleSkeleton = ({
   const isSharingSavingEnabledByTPA = savingEnabled || sharingEnabled;
   const isSharingSavingEnabled =
     isSharingSavingEnabledByTPA && isSharingSavingEnabledExternal;
-  const domainSpecificUrl = hostName
-    ? getDomainSpecificUrl(hostName, "https://www.thetimes.co.uk")
-    : "https://www.thetimes.co.uk";
+  const domainSpecificUrl =
+    hostName && getDomainSpecificUrl
+      ? getDomainSpecificUrl(hostName, "https://www.thetimes.co.uk")
+      : "https://www.thetimes.co.uk";
 
   const isLiveOrBreaking = getIsLiveOrBreakingFlag(expirableFlags);
   const [polygonUrl, setPolygonUrl] = useState([]);