+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.128.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.128.0...@times-components/article-skeleton@1.128.1) (2024-03-25)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 # [1.128.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.127.0...@times-components/article-skeleton@1.128.0) (2024-03-19)